pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "codest", email = "thecodest@ya.ru"},
]
dependencies = [
    "Django",
    "djangorestframework",
    "psycopg[binary]",
    "celery[redis]>=5.3.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}


[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
